char: move qemu_openpty_raw from util/ to char/
It is only needed by char-pty.

Fix the code style while at it.

Message-Id: <20220323155743.1585078-6-marcandre.lureau@redhat.com>
